Add compatibility with Zidoo Native Player
Add compatibility with Zidoo Player, you have to enable external player and direct path option in the app settings. You also have to check the zidoo player option and add smb username and password if needed. I have tested it on the Zidoo Z9X in the v6.2.0_g firmware.